Use Response module instead of Xander.CBOR
Similar to how Xander.Query works.
